+++ treasuryArbitrage/src/edu/nju/treasuryArbitrage/network/DataInterfaceToServer.java
@@ -186,7 +186,7 @@ public class DataInterfaceToServer implements DataInterface {
 	@Override
 	public boolean Order(String More_contract,
 			String Blank_contract, double more_price, double blank_price,
-			int hand, int guarantee) {
+			int hand,double guarantee) {
 		// TODO 自动生成的方法存根
 		UserVO user = userbl.getUser();
 		return tradebl.order(user.getUserID(), More_contract, Blank_contract,
